+++ packages/pluggableWidgets/combobox-web/src/__tests__/StaticSelection.spec.tsx
@@ -78,8 +78,7 @@ describe("Combo box (Static values)", () => {
             selectedItemsSorting: "none",
             customEditability: "default",
             customEditabilityExpression: dynamic(false),
-            filterInputDebounceInterval: 200,
-            datasourceFilterDebounceInterval: 200
+            filterInputDebounceInterval: 200
         };
         if (defaultProps.optionsSourceAssociationCaptionType === "expression") {
             defaultProps.optionsSourceAssociationCaptionExpression!.get = i => dynamic(`${i.id}`);